Bump versions for Play 2.3.2; offlineTests not working
 Play to 2.3.2
 sbt-remote-control to 1.0-8756df5dedb13e490ab9ce8a11dbb5a0b5c02362
 sbt-echo to 0.1.1.3 (which appears to be newer than 0.1.3 - not sure
   what happened there)
 Added another junit version, plus some explicit plexus and jsch deps.
 Do not understand why any of these were suddenly necessary.

In this commit things are broken; it seems that jsch never gets
downloaded and is not in local-repository after the local-repository-created
task. Then the offlineTests can't find it.

Some of the offlineTests failures are just because templates haven't
been updated to latest Play, but play-scala and play-java are somehow
failing on this jsch thing.